Property traits: descriptor fundamentals
This is the refactor of the descriptor code which started this
fork.  Admittedly, the use of MetaProperty to allow the Property()
constructor to return a different type to the one requested is a
gimmick; it could equally be a simple function.  However I wanted to
keep the general feel of the API and this allows the constructor symbol
and the class symbol to be the same.  It's also a demonstration of the
power of python's metamodel.