Add true Lazy properties
Python's descriptors API describes properties that have a __get__ method
which sets the instance's property slot (ie, .__dict__[x]), and allows
subsequent attribute accesses to skip the descriptor.  This seems
desirable, so add it.  It necessitates not deriving the 'property' type,
because that type defines a __set__ method which would interfere with
the mechanism.  It also requires making read-only properties the special
case (because to implement that, they must also define __set__).