Select Property subtype based on options
Having to import the exact Property type you want is a bit fiddly.
There is already a mechanism for the constructor to select a sub-class,
and currently passing unknown constructor arguments is a hard error.
Instead, try to automatically select extra traits to add based on which
class the unknown constructor argument was defined in.