Add 'primary key' concept
This will be used for comparing sets of objects in the diff mechanism.
If the primary keys compare equal, the items are considered the same
item but possibly with changed attributes.  For now, just 'expose' the
primary key properties using the __str__ method.  Implement __repr__
along the way, for posterity.