auto-vivifing setter for FieldSelector
When taking specific differences and applying them to the previous data
structure, it is important to be able to create objects as required
rather than just set attributes in existing objects.

now there is a problem: this means the FieldSelector has to create
correctly typed python objects along the way.  When it comes to
collection types, this is even harder: it needs to create a Collection
object with the correct item type set.  This poses a problem, because
the FieldSelector would have to know this argument during construction.
But at the moment, FieldSelector objects are not tied to the data types
they reference.

So, replace the old Collection type which used an instance property with
a type in it, with an implementation of higher order types/generics.

Yes, really.

It works, and you can see that it actually simplifies the
implementations of the *CollectionProperty classes.  I wouldn't have
done it if it wasn't so few lines of code, but I'm still probably going
to hell for it.

Oh, I called the auto-vivifying put 'post', to continue the REST
metaphor of FieldSelector.