Implement MultiFieldSelector to 'filter' data structures
Sometimes when sending objects around, a subset of the data in the full
object is desired.  One way to represent the allowed data is to use
FieldSelectors, and their combination represents the subset to return.
Implement this; note that the MultiFieldSelector.get operation returns
a data structure, not just the fields referenced.