diff: permit duck-typed diffing against 'alien' types
Sometimes you want to compare normalize objects which are not normalize
classes, or perhaps you don't care about subclasses.  Add a new option
to diff (via DiffOptions) which enables "duck type" diffing; that is,
don't check that the object *is* the same, just that it *behaves* the
same way; ie, returns the same values from its data accessors (or
"quacks" the same).