diff: aliens have identity, too
When comparing an "alien" class against a Record, the concept of
identity needs to be carried over, too.  Factor out the two places where
the previously trivial function to extract a primary key tuple into a
single function.

This gets complicated if one part of the primary key is a Record.  This
is not a common concept to most RDBMS users; in fact even Postgres
couldn't use compound types as a primary key for some time.  In order to
compare the primary key identity between alien and native types, the
primary key must first be recursively reduced to (hashable) "atomic"
values.

So, the "record_id" function becomes another "visitor pattern" function,
built to navigate the type hierarchy on its own, without relying on
type().

Avoid returning the object itself when no primary key is defined;
instead, try to reduce the object to hashable things.