diff: show identity of record on str(record)
The logic in the stringification function slightly overlapped that
of the primary key function.  And in any case, it is a useful way
to informally distinguish objects.  Without a primary key defined,
the str() was falling back to repr(), which was very verbose.  Use
the new tuple reduction function instead.