Improve representation of richenum values
The previous mapping of richenum to a property was clumsy.  The
attribute really wants a type, but OrderedRichEnumValue was not a very
good answer to that.  Instead, it used an int sub-type (a type-checked
int).  This meant you couldn't compare the value to the richenum class
attributes as the module wants you to.  RichEnum can't use a common base
class for enum values of a particular enum because of the order of
construction of the objects, but this simple trick: a simple empty base
class, can solve the problem.