Fold 'ListRecord' and make_generic into one
This ListRecord class deriving list was clever in a sense, but a bit of
a parlor trick.  It can be replaced by ListCollection with surprisingly
little refactoring, and any list functionality can always be rebuilt
using dunder-methods etc.  Along the way, get rid of 'record_cls' and
require the downstream consumers to use 'itemtype'.

There's still a slight schism: types created by ListProperty(ItemType)
will not be the same type as those created using class
ItemTypeList(RecordList):\n\t itemtype = ItemType.  This is a bug, as is
the global GENERIC_TYPES, which tries to solve this problem for multiple
users of ListProperty(Item).  The fact it is global breaks the test (fix
by renaming Item to Eyetem in one test), and it should be able to find
any *List classes defined in the regular way.