make_generic: solve collection property type schism
So, there were a couple of problems with the way generic types were cached.
They were cached by the generated name of the class, which was not unique
between modules.  Secondly, it was not possible to re-use a collection in a
ListProperty by passing the collection type as an isa= parameter; detect
and validate this situation and skip generating the class in that case.