Support 'None' as a value type
If you specify default=None, this is somewhat different from not specifying
a default.  Use a different default to the __init__ constructor, and allow
None to be a value.  As a special case, if a property is specified as
required=True, then don't allow the value to be set to None if no type is
specified via isa=.