JSON marshalling: test (and fix) JsonProperty xform funcs
The json_in and json_out arguments to the JsonProperty constructor are
convenient ways to provide transform functions on the way in and out.  They
weren't being called.  Also, the distinction between a property which is
not set and one which is set to None was not clear.  The Property.to_json()
method changes to accept the slot value, not the instance (as if it were a
descriptor __get__)