JSON marshalling: split dual-functionality of 'from_json' into two
from_json behaved slightly differently depending on how it was called, but
this caused problems when specializing behavior.  This refactoring breaks
the JSON-to-init-arguments part out separately from the actual instance
construction.  Along the way, Collection became a Record, resulting in some
restructuring of the module.  Somehow, all the tests still pass.  This may
reflect a lack of test coverage... :-)