diff: remove a goto
Collections are now Records, so items may match multiple functions in
this dict.  Put the actual diff inside the loop, so that it compares
both the properties and the collection values for Collection types.

Solves this occasional error:

    FAIL: Test that all the pieces work together
    ----------------------------------------------------------------------
    Traceback (most recent call last):
      File ".../normalize/tests/test_compare.py", line 345, in test_complex_objects
        compare_record_iter(wall_one, wall_two), expected_differences,
      File ".../normalize/tests/test_compare.py", line 77, in assertDifferences
        set("<DiffInfo: %s>" % x for x in expected)
    AssertionError: Items in the second set but not the first:
    '<DiffInfo: REMOVED .posts[0].comments[0]>'
    '<DiffInfo: MODIFIED .posts[0].edited>'
    '<DiffInfo: REMOVED .posts[0].comments[1].poster.interests[1]>'
    '<DiffInfo: MODIFIED .posts[0].content>'
    "<DiffInfo: ADDED .posts[0].comments[1].poster.info['birth name']>"
    '<DiffInfo: ADDED .posts[0].comments[2]>

    ----------------------------------------------------------------------

(about the subject: statements like 'break', or 'return' before the end
of the function, and so act as a "euphamism that “goes to” without
saying **go to**", to quote Knuth (1974))