Add a diff option so that empty slots can == values set to '' or None
When comparing data that comes out of NoSQL blob stores, often it is
equivalent to have '', None or the key not existing as equivalents.  In
principle, the normalize Record should perform this normalization based on
the semantics of the data source, however being able to suppress such
differences on diff is a convenient short-cut when dealing with partially
normalized data.