Fix duck type diffing and ignore_empty_slots
If you were diffing, the new feature to see if the other slot was "empty" would
end up comparing a collection to a string, and throw:

      File ".../normalize/diff.py", line 151, in compare_record_iter
        propval_a = options.normalize_val(getattr(a, propname, Nothing))
      File ".../normalize/diff.py", line 120, in normalize_val
        elif self.ignore_empty_slots and value in ("", None):
      File ".../normalize/coll.py", line 38, in __eq__
        return self.itemtype == other.itemtype and self.values == other.value
    AttributeError: 'str' object has no attribute 'itemtype'

Fix this in two ways: the __eq__ operator should not be that fragile;
fix it.  Also, don't compare to a string: an 'empty' slot really must be
None or the empty string.  Put that in a new method to make it more
easily customizable if someone else wants to do so, and make it more
specific, as I think is a good reasonable default.