coll: split init_values into two class methods and a property
In order to support constructor visitor patterns (eg, 'from_json'), the
visitor class should be able to "map" over input which is not already
instances of the visited types.  In order to do this, some logic which is
buried inside 'init_values' needs to be used independently.  So, split the
various functions into class methods that can be called without instances:
coll_to_tuples takes a collection and returns the itertuples() form of a
collection, and tuples_to_coll() takes a generator and returns a correct
collection, with items coerced into the correct type.  Adjust the visitor
to use these new methods, removing an ugly type switch.

Necessarily, the API for coercing from unusual collection formats changes,
see the change in test_property for how that impacts the downstream.