doc: autodoc the remaining classes
There are still a lot of empty headings, and the rest of the reference
manual needs tidying up.  But this quick pass has at least stubs for
almost everything.  Quite a few internal classes got renamed to be private
and there was one more "more nothing than nothing" class which got a
shonky __repr__ method.