Automatically mix Property types together as needed
When creating a new Property type via make_property_type, it was
necessary to choose the "computed" Property sub-type.  This is because
the Property() constructor was not able to create subclasses.  Also, I
was not sure that blindly mixing together Property subclasses without
providing resolution methods would work in the general case.  However,
after much work with it, the task of manually creating all of these
combinations of Property sub-classes (which will, in general, simply use
'pass' as the implementation) completely sucked.

So, write a function to create combinations of Property types, as they
are declared.  At the stage where we would have thrown an exception
before, try to call it.  Remove most of the old empty classes that
worked around this feature not being there.  What could possibly go
wrong?

Also, 'lazy' traits became safe by default, and the 'slow' trait was
killed.