Only select SafeProperty IF a type constraint exists
Previously, the only way to avoid getting the 'safe' trait was to
pass ``traits=['unsafe']`` to the Property() constructor.  This was a
suitably guarded safety latch, however if the user of the property
didn't pass any kind of type constraint (eg, required, isa, check),
then there was no need to use safe.  Make it check, and also update the
tests where they needlessly cared.