Allow 'coerce' function to return 'None' to indicate empty slot
Often coerce functions which are supposed to silently swallow values (ie,
without raising an exception) will do this by returning None.  Allow this
to result in the slot being left empty on object initialization, but not on
assignment (ie, assigning a value which coerces to 'None' is not a valid
alternative to del obj.attribute).