Apply diff options to 'record_id' function
When comparing filtered objects, declared primary keys can become
invalidated.  This ends up surfacing differences between "" and None and
not set, when membership in collections falls back to using the entire
value of members.  There are effectively still two sections of code which
deal with comparison: one is the recursive functions: compare_record_iter
etc, and the other is the collection comparison function used in
compare_collection_iter.  The latter is essentially an optimization, so
that hashing can be used.  The "correct" alternative would be an O(N²) loop
which diff's members in the collections with other members until it finds
ones which are the same.  Instead of doing this, it's possible to pass the
value reduction functions into record_id(), so let's do that.