Make ignore_empty_slots also ignore 'empty' records
If you were comparing two records, and one side had no value in a slot but
the other side had a record which was all effectively empty, it would still
show as added/removed, even with 'ignore_empty_slots'.  This change expands
that so that if that option is set, it continues to recurse, but passing in
the ``_Nothing`` instance and only surfaces an added/removed event if one
comes out of the descent.  This is better than trying to implement this in
the normalization hook (ie, normalize_slot), because it can be filter-aware
and re-use the existing descent.  This probably isn't a complete approach:
notably, the logic does not work for items in a collection.  It's a start.