Avoid emitting {'values':[...]} on JsonRecordList classes
JsonRecordList classes have an extra extraneous property,
unknown_json_keys, which somewhat ironically is impossible to get anything
into by default, and is not even visited without setting ``extraneous``.
However this default reducer was making the mapped result more complex than
it would typically need to be.  Instead, decide based on whether any of
those properties are filtered or not whether to use the complex form.