Fix Property metaclass magic for multiple-trait property types
Duck-typing property types implies *all* of the traits in the property
class, not just the one defined in that class.  This was preventing its
use for types like ListProperty, which has two traits: ``coll`` and
``list``.  Fix it by storing the full 'traits' tuple in DUCKWARGS and
adding that to the combined list of traits for the property being
created.