LazyProperty should double-check before firing
LazyProperty expects that Python itself will short-cut its read when it
is not called as SafeLazyProperty; however, if you are using a
metaprogramming technique which is calling __get__ directly to get the
current value of the attribute, the lazy attribute will fire anyway.  To
avoid this, double check that the attribute is not already set before
firing.