Handle mixed presence of fields in FieldSelector.delete and .get
If a field selector refers to a set, such as ``["foo", None, "bar"]`` (or
``.foo[*].bar`` in the path form), then ``delete`` and ``get`` would fail
on the first unsuccessful lookup.  Instead, wrap the iteration in a try:
except: block, and only raise the exceptions if at least one exception was
raised and no values succeeded; otherwise continue.