MultiFieldSelector: defining ambiguities around collection-like methods
Code which used this functionality for things like filtering discovered
an abiguity: the FieldSelector IN MultiFieldSelector has an edge case
where the FieldSelector is short.  Some users want to proceed (eg, with
visiting a type structure) if ANY fields are defined at that level;
others (asking, for instance, "is this field completely included in this
field selector?") should only proceed if ALL fields are defined at that
level.

For instance, ("foo",) in MultiFieldSelector(("foo", "bar")) would
return true, but if you called fs.get(mfs.get(obj)), it would not return
all of the fields that fs.get(obj) would.  The FieldSelector refers to a
set of fields, and the set referred to by the MultiFieldSelector is not
a superset.  But code referring to the MultiFieldSelector to know
whether a field is filtered needs to know that there is something under
"foo".

So, allow the subscript operation to return the sub-MultiFieldSelector.
In this mode if nothing exists, it won't throw KeyError but instead just
return None.  Document the behavior and adjust the included visitors and
methods to call the new functions.