Allow bool(Record) to determine NULL-ness
This was already the case with collections due to default rules.  This
test allows you to check that a record has at least one defined,
non-extraneous attribute using bool(Record).  This was already the case
for collections: empty collections were previously false because the
default python rule would fall back to __iter__, and return false if it
returned nothing.  Allow this to continue to work, by logically or'ing
the length of the collection and the emptiness of the record.