JsonProperty: json_out return value should not be converted
It was already an explicit API for ``json_in`` to expect the user to
recurse into ``from_json`` if required, so it was strange that the
``json_out`` return value was unconditionally passed to ``to_json`` for
a second marshal step.

Generally ``json_out`` would be used to return plain JSON values such as
strings, and fully marshalled (``json.dumps``-ready) dictionaries, so
this pass would normally have essentially been a no-op.  This new
behavior allows the method to return values which are NOT processed via
``to_json`` (eg, if the result is being passed to ``json.dumps`` with a
custom value marshaling class, this will allow those values to be
passed through as-is).  It is also faster and uses less memory to not
recursively copy all the values which are passed through.  The default
case where no ``json_out`` trait is defined will still use ``to_json``.

It is an API change, and while it's a bit rotten to sneak this in during
a minor point release, there's been a recent major release and the old
behavior wasn't documented anyway.