Rework semantics for conversion in JsonProperty.json_data
The tweak to to_json semantics was severely broken: *all* JsonProperty
properties had 'to_json' defined, not just ones which declared
``json_out``.  Previously the expectation of that method was that it
would have 'normalize.record.json.to_json' called on the return value.
However this conversion is now considered optional if you declare
``json_out``.  JsonProperty.to_json() needs to get slightly more
sophisticated, accepting a function to continue with if the property is
not defined.