Make empty attributes return magic falsy objects
Making 'foo.attribute0' return an instance of Record if 'attribute' was
set isa=Record, and making Record falsy if all its attributes are falsy,
was probably a mistake.  In practice many consumers end up doing the
construct listed in the release notes for 0.7.0.  Replace it with an
alternative which still checks for bad spelling on attributes but does
not have this behavioral side effect.