Add a new diff option to return 'moved' items
Currently, as far as diff is concerned, collections are regarded as bags of
things, where the things are the important things and the keys are regarded
as trivial properties only useful for referring to items in the diff
output.

With dictionary collections, the keys are likely to indicate a stronger
notion of identity.  It might actually be a significant difference that
two values have swapped keys; less so for lists.  Add a new diff option to
allow these differences to be returned, too.