Fix Visitor and Collections with non-record Items
The 'unpack' stage here needs to handle the case where it's been given a
non-Record type; typically this arises during processing of collections
where the item type is not a Record.  Alter the Visitor sub-class API
slightly to allow these stages to return 'None' for the properties
stage, indicating that 'apply' should be called directly and there is no
property.