Visitor: don't skip 'reduce' for collections
For the typical case where collections had no extra properties, reduce
was skipped on cast.  This meant that the collections were not converted
to normalize collections; mainly noticable if used with eg JsonProperty
values with a json_name set, due to implicit casting.  Or as this new
test case demonstrates, casting directly to a simple collection of
records.  Oops!