Add a helper for creating subtypes
As the subtype test in tests/test_types.py demonstrated, the syntax for
creating types which refer to subsets of other types was complicated and
required the creation of metaclasses.  Add a convenience function for
creating these types which just expects a base class and the two key
functions: is the value in the set of allowed values, and how to convert
other values to allowed values.