Allow sub-type coerce functions to return None to squash attributes
This works for regular coerce functions, so should probably also work
for subtype coerce functions.  Using a subtype's constructor to cast
will still throw an exception if it didn't work, but this exception can
be caught and the value retrieved, just to see if it is None or not.
The error raised will still be more or less equivalent.