exc: make CoercionError errors more consistent
In order to handle parse exceptions more gracefully, make the coercion
errors that most bad input will encounter more consistent.  Move to
'passed' for most input values, as well as 'recordtype', 'colltype' and
'itemtype', 'subtype' for the various type objects which are available.
Also pass the actual objects in the exceptions, not stringifications.