JsonRecordList & Dict: pass-through already-constructed values
When assigning to a property containing a collection of records, the plain list
or dict will be upgraded to the collection type, using the coerce function.
However, the expectation of the Json* classes that the first argument is JSON
unless passed as keyword arguments overlaps with this dual function as a
general coercion API.  To fix, detect items in the input which are already
blessed python objects, and skip the Json conversion stage.

Without the fix, the new test fails with:

    ======================================================================
    ERROR: test_json_coll (test_coll.TestCollections)
    ----------------------------------------------------------------------
    Traceback (most recent call last):
      File ".../tests/test_coll.py", line 272, in test_json_coll
        jdr.item_map = {"bill": jlr.item_list[0]}
      File ".../normalize/property/coll.py", line 66, in __set__
        obj, self.type_safe_value(value),
      File ".../normalize/property/__init__.py", line 228, in type_safe_value
        self.valuetype.__name__
    CoerceError: coerce to ItemMap for JDR.item_map failed with value {'bill': Item(name='Bob')}: ItemMap constructor raised: Json conversion failed at .bill; Cannot interpret Item(name='Bob') as a Item constructor

    ----------------------------------------------------------------------